Migrated to Swift 5.9, Dropped SwiftToolsSupport, Updated Swift argument parser
