+++ .swift-version
@@ -1,1 +1,1 @@
-5.0
+5.9
