+++ Makefile
@@ -27,7 +27,7 @@ REPODIR = $(shell pwd)
 BUILDDIR = $(REPODIR)/.build
 
 EXECUTABLE_DIRECTORY = ./.build/x86_64-apple-macosx/debug
-TEST_RESOURCES_DIRECTORY = ./.build/x86_64-apple-macosx/debug/xcprojectlintPackageTests.xctest/Contents/Resources/
+TEST_RESOURCES_DIRECTORY = ./.build/x86_64-apple-macosx/debug/XCProjectLintFrameworkTests.xctest/Contents/Resources/
 
 error:
 	@echo "Please choose one of the following targets: build, clean, test, xcode"
