+++ Sources/XCProjectLintFramework/CheckForWhitespaceSpecifications.swift
