+++ Tests/XCProjectLintFrameworkTests/TestData/Bad.xcodeproj/project.xcworkspace/contents.xcworkspacedata
