+++ Tests/XCProjectLintFrameworkTests/TestData/Bad.xcodeproj/project.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
