+++ Tests/XCProjectLintFrameworkTests/TestData/Bad/main.swift
