+++ Tests/XCProjectLintFrameworkTests/TestData/Good.xcodeproj/project.xcworkspace/contents.xcworkspacedata
