+++ Tests/XCProjectLintFrameworkTests/TestData/Good.xcodeproj/project.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
