+++ Tests/XCProjectLintFrameworkTests/TestData/Good/AppDelegate.swift
@@ -14,13 +14,13 @@
 
 import UIKit
 
-@UIApplicationMain
+@main
 class AppDelegate: UIResponder, UIApplicationDelegate {
   var window: UIWindow?
 
   func application(_: UIApplication, didFinishLaunchingWithOptions _: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
     // Override point for customization after application launch.
-    return true
+    true
   }
 
   func applicationWillResignActive(_: UIApplication) {
