+++ Tests/XCProjectLintFrameworkTests/TestData/Good/Assets.xcassets/AppIcon.appiconset/Contents.json
