+++ Tests/XCProjectLintFrameworkTests/TestData/Good/Assets.xcassets/Contents.json
