+++ Tests/XCProjectLintFrameworkTests/TestData/Good/Base.lproj/LaunchScreen.storyboard
