+++ Tests/XCProjectLintFrameworkTests/TestData/Good/Base.lproj/Main.storyboard
