+++ Tests/XCProjectLintFrameworkTests/TestData/Good/main.swift
