+++ Tests/XCProjectLintFrameworkTests/TestData/Good/xcconfigs/Common.xcconfig
