+++ Tests/XCProjectLintFrameworkTests/TestData/Good/xcconfigs/Debug.xcconfig
