+++ Tests/XCProjectLintFrameworkTests/TestData/Good/xcconfigs/Release.xcconfig
